# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Faydamed API Authentication",
        "description": "Authentication API endpoints for Faydamed system",
        "contact": {
            "email": "support@faydamed.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "https://faydaapi.dotprogrammers.com",
            "description": "Development server"
        }
    ],
    "paths": {
        "/api/api-credentials": {
            "get": {
                "tags": [
                    "API Credentials"
                ],
                "summary": "List all API credentials",
                "description": "Retrieve a list of configured API providers (Google Vision, DocuSign, etc.)",
                "operationId": "listApiCredentials",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "API Credentials"
                ],
                "summary": "Create/Update API credential",
                "description": "Register a new API key for a provider",
                "operationId": "storeApiCredential",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "provider",
                                    "name",
                                    "key"
                                ],
                                "properties": {
                                    "provider": {
                                        "type": "string",
                                        "example": "google_vision"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "API_KEY"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "AIza..."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Credential saved"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/api-credentials/{id}": {
            "get": {
                "tags": [
                    "API Credentials"
                ],
                "summary": "Get credential details",
                "operationId": "getApiCredentialById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Credential not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "API Credentials"
                ],
                "summary": "Update API credential",
                "operationId": "updateApiCredential",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Credential updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "API Credentials"
                ],
                "summary": "Delete API credential",
                "operationId": "deleteApiCredential",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credential deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}